refactor(match_history): split MatchHistoryFrame into mixin package
Convert widgets/frames/match_history.py (650 LOC single file) into a
package with responsibilities split across:

- __init__.py: public API (MatchHistoryFrame, main)
- frame.py: class definition, __init__, and _build_ui (UI construction)
- handlers.py: MatchHistoryHandlersMixin — tree events (on_item_*),
  worker callbacks (refresh_history, _populate_history,
  _handle_history_error, _init_username), status/stat mutators
  (_set_busy, _update_metrics, _update_opp_stats, _clear_opp_stats),
  on_close
- properties.py: MatchHistoryPropertiesMixin — i18n (_t),
  username accessor (_set_username, _get_opponent_name), pure-data
  helpers (_iter_matches, _parse_date, _parse_date_input,
  _within_range)

The class composition becomes:
  class MatchHistoryFrame(HandlersMixin, PropertiesMixin, wx.Frame): ...
so external imports (``from widgets.frames.match_history import
MatchHistoryFrame``) continue to work unchanged via the package
re-export. pyproject.toml's E402 per-file-ignore now points at
widgets/frames/match_history/frame.py where the sys.path bootstrap
lives.

This is the first widget refactored to the target "package of mixins"
layout described in issue #427; remaining frames/panels follow the
same template in subsequent commits.

Refs #427 (second pass, partial).
