+++ widgets/frames/match_history/__init__.py
@@ -1,0 +1,7 @@
+"""Match history viewer package."""
+
+from __future__ import annotations
+
+from widgets.frames.match_history.frame import MatchHistoryFrame, main
+
+__all__ = ["MatchHistoryFrame", "main"]
